Chore: rename apps to applications for clarity
