+++ api/src/app/controllers/loanApplicationsController.ts
@@ -1,13 +1,13 @@
 import { Request, Response } from "express";
-import { calcLoanOffers } from "../services/loanAppsService";
-import LoanApp from "../models/loanApp";
+import { calcLoanOffers } from "../services/loanApplicationsService";
+import LoanApplication from "../models/loanApplication";
 
-export const createLoanApp = (req: Request, res: Response) => {
+export const createLoanApplication = (req: Request, res: Response) => {
   const loanFormJson = req.body;
   const { price, deposit, loanTerm } = loanFormJson;
 
   // Save the application to be processed by other parts of the app
-  LoanApp.createApp(req.body);
+  LoanApplication.createApplication(req.body);
   const response = calcLoanOffers(price, deposit, loanTerm);
 
   res.status(201).json(response);
