+++ api/src/app/routes/loanApplicationsRoutes.ts
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import { createLoanApplication } from "../controllers/loanApplicationsController";
+
+const router = Router();
+
+router.post("/loan-applications", createLoanApplication);
+
+export default router;
